+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.6 (unreleased)
+
+- Return error if attach/detach action fails
+
 ## 0.1.5
 
 - Fix empty tunnel-interface config of `sdwan_cisco_vpn_interface_feature_template`