+++ docs/guides/changelog.md
@@ -7,6 +7,10 @@ description: |-
 
 # Changelog
 
+## 0.1.6 (unreleased)
+
+- Return error if attach/detach action fails
+
 ## 0.1.5
 
 - Fix empty tunnel-interface config of `sdwan_cisco_vpn_interface_feature_template`